Fix NullPointerException
The NullPointerException would occur if pathTo() is called with a move
from the cache. From the passed move to the current move a list will be
constructed in reverse order, forming the path from the current move to
the best move. The loop would throw a NullPointerException if the move
was not a (grand)child move of the current move, which can happen if the
move comes from the cache. The cache holds the best move resulting from
a certain position, but the moves leading to that position can differ,
which makes it impossible to create the path to the best move.