Allow function instead of consumer in evaluator
In the old situation only a consumer was allowed as parameter. The
consumer would perform setup the board, evaluate a list with one move
and then return the value the move received during the evaluation.

In the new situation, a function can be passed in that will be called
with the setup board and the move. The function then returns the value
of the move.