Reimplemented AiController
Let actions be executed sequentially, so that updating the status does
not need to be synchronized. Schedule a tick action every second to be
able to change the maximum depth delta in case quiescence search takes
too long.
Still need to tweak the algorithm cause thinking often takes twice the
time it was given.