Improved time limit
Use ratio of percentage of remaining time and percentage of remaining
amount of moves to search to determine the max depth delta. If the
maximum time has passed, then the max depth delta is increased every
step and every seocnd to force the thinking to finish as soon as
possible.