Implement redo move function
It is now possible to undo as far as the beginning of the game and redo
moves. The currently selected move is shown bold in the moves panel.
When undoing moves, both players will be controlled by humans as a way
to pause the computer from thinking up new moves again. If instead of
redoing moves a new move is made (either by a human or computer player),
all moves following the current move are removed and the new move is
added as last move.