Do not change value in case of alpha/beta cut-off
Changing the value in case of alpha/beta cut-off resulted in move values
that contained the alpha/beta value. Not changing the move value will
keep real values in the moves that will never be selected as best move.
(That is the whole point of applying alpha/beta cut-off). Using the
alpha/beta value sometimes let those cut-off moves be selected.