Clear expected opponent move after undoing move
Imagine the following scenario: human plays against a computer player.
A couple of moves have been played. The computer player has stored an
expected next move and the response move to the expected move. The human
player undoes a couple of moves and then plays the move that was
expected by the computer. The problem is that the expected move was
expected in a different position than the current position. The computer
playes the stored response move, which changes the board in an illegal
way.

This problem has been fixed by clearing the expected move and its
response as soon as the user undoes or redoes a move.