Discourage Rh1-g1 and Rh8-g8 in opening
Turns out that getPieceMovingFromKingSideValue() caused the rook on the
king side move one square. Although not explicitly mentioned in the
article about the Technology Chess Program, I guess that this move
should not be rewarded for being a piece moving from king side. I guess
that moving a piece from king side should move away the knight and
bishop to allow castling. The rook should therefore not be moved for
this reason.